requirements.txt
fastapi
uvicorn
httpx
starlette
python-dotenv
websockets
